requirements.txt
discord.py==1.7.3
asyncio==3.4.3
pymongo==4.1.1
dnspython==2.2.1